app/requirements.txt
flask
flask-restful
python-keycloak
minio
flask-jwt-extended
